+++ scripts/temp_archive/align_NPRW_BR.py
@@ -125,10 +125,10 @@ def load_intan_adc_2ch(npz_path: Path) -> tuple[np.ndarray, np.ndarray, float]:
         for k in keys:
             a = z[k]
             if a.ndim == 2 and a.shape[1] >= 2:
-                ch0.append(a[:,0].astype(np.float64))
-                ch1.append(a[:,1].astype(np.float64))
+                ch0.append(a[:,0].astype(np.float32))
+                ch1.append(a[:,1].astype(np.float32))
             elif a.ndim == 1:
-                ch0.append(a.astype(np.float64))
+                ch0.append(a.astype(np.float32))
         if not ch0:
             raise RuntimeError("No usable chunk_* arrays in ADC NPZ")
         tri = np.concatenate(ch0)
